#80 refactor: KakaoPay Payment Failure View Refactoring
