+++ src/views/OrderHistoriesView.vue
@@ -6,8 +6,6 @@
 import OrderHistory from '@/components/store/order/OrderHistory.vue';
 import { defineComponent } from 'vue';
 
-
-
 export default defineComponent({
     components: {
     OrderHistory
