+++ src/views/PosView.vue
@@ -6,13 +6,9 @@
 import HomeBoard from '@/components/store/common/HomeBoard.vue';
 import { defineComponent } from 'vue';
 
-
 export default defineComponent({
     components:{
         HomeBoard
     }
 })
 </script>
-
-<style lang="scss" scoped>
-</style>
