+++ src/views/SignUpView.vue
@@ -7,14 +7,8 @@ import { defineComponent } from 'vue';
 import SignUpForm from '../components/user/SignUpForm.vue'
 
 export default defineComponent({
-
     components: {
         SignUpForm
     }
-
 })
-
 </script>
-
-<style>
-</style>
