--- src/views/kakaopay/KakaoPayFail.vue
@@ -1,27 +1,0 @@
-<template>
-    <div class="view">
-        결제에 실패하였습니다.
-    </div>
-    
-</template>
-
-<script lang="ts">
-import { defineComponent } from 'vue';
-
-
-export default defineComponent({
-
-})
-</script>
-
-<style lang="scss" scoped>
-
-.view {
-    position: fixed;
-    top: 40vh;
-    width: 100%;
-    color: yellow;
-    text-align: center;
-}
-
-</style>
